remove settings code
