+++ remme/certificate/client.py
@@ -87,7 +87,7 @@ class CertificateClient(BasicClient):
 
     def revoke_certificate(self, crt_address):
         payload = self.get_revoke_payload(crt_address)
-        return self._send_transaction(CertificateMethod.REVOKE, payload, [crt_address])
+        self._send_transaction(CertificateMethod.REVOKE, payload, [crt_address])
 
     def get_signer_pubkey(self):
         return self._signer.get_public_key().as_hex()
